fix: sb

+++ Assets/Script/Misc/Types/Net/DLProgress.cs
@@ -3,7 +3,7 @@ namespace MajdataPlay.Net
 {
     public readonly struct DLProgress
     {
-        public double Length { get; init; }
+        public long Length { get; init; }
         public double Progress { get; init; }
     }
 }
